requirements.txt
google-generativeai>=0.8.0
pydub>=0.25.1
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.26.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
